move logic to remote function
